Split query tests to separate file to stay under 1000 line limit
The traits.rs file was exceeding the 1000 line limit after adding
comprehensive query pattern tests. This commit splits the tests:

- traits.rs: Original trait tests (991 lines)
- query_tests.rs: New query pattern tests (913 lines)

Both files now comply with the file size limit check.
